More improvements and ToDos
